Add method selector to disable G keys
Some keyboards require the disableGKeys buffer to be sent via
hidapi_write, but others (e.g. Logitech G710+) need the buffer to be
sent via hidapi_send_feature_report. This commit adds a flag to the
KeyboardInterface class that indicates whether to use the write or the
send_feature_report method, and switches on that accordingly.